I started on a different repo, but I've since moved to this fork.
Program: makes a grid and places the starting room in the centre, and initialises the player and game.

Game: has a loop to ask the user what to do and calls methods.

Player: sets up the player, has methods to add items to room, display info, moving rooms, picking up items and displaying the map

Room: sets up the room, has methods to get the description, add items to room, add enemy to room and add exits to room.

GameData: stores information that can then be given out randomly.