Changes from code reviews
Implemented some changes according to the reviews received from my peers. I've changed the game code heavily so that it doesn't look as cluttered as it used to. The name setter works properly now and asks the user to continue inputting a name until it is valid. Changed how items work to make it a bit easier to implement items in the future. Added comments to make the code easier to understand.